refactor: 영화 데이터 및 검색 함수를 객체로 관리하도록 변경
