+++ app/src/main/java/woowacourse/movie/presenter/MovieListPresenter.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.presenter
 
-import woowacourse.movie.MOVIES
 import woowacourse.movie.adapter.MovieAdapter
+import woowacourse.movie.model.MovieData.MOVIES
 import woowacourse.movie.presenter.contract.MovieListContract
 
 class MovieListPresenter(private val movieListContractView: MovieListContract.View) : MovieListContract.Presenter {
