+++ app/src/main/java/woowacourse/movie/presenter/TicketingPresenter.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.presenter
 
-import woowacourse.movie.findMovieById
 import woowacourse.movie.model.Count
+import woowacourse.movie.model.MovieData.findMovieById
 import woowacourse.movie.model.Result
 import woowacourse.movie.model.Tickets
 import woowacourse.movie.presenter.contract.TicketingContract
